+++ docs/conf.py
@@ -11,7 +11,8 @@ html_theme = 'sphinx_book_theme'
 exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
 html_static_path = ['_static']
 html_theme_options = dict(
-    extra_footer='Theme by the <a href="https://ebp.jupyterbook.org">Executable Book Project</a>.',
+    extra_footer='Theme by the <a href="https://ebp.jupyterbook.org">'
+    + 'Executable Book Project</a>.',
     repository_url='https://github.com/pauleveritt/viewdom',
     use_repository_button=True,
 )