+++ examples/index/conditional.py
@@ -8,10 +8,12 @@ message = 'Say Howdy'
 not_message = 'So Sad'
 show_message = True
 result = render(
-    html('''
+    html(
+        '''
     <h1>Show?</h1>
     {message if show_message else not_message}
-''')
+'''
+    )
 )
 expected = '<h1>Show?</h1>Say Howdy'