+++ examples/usage/loopingA.py
@@ -4,11 +4,13 @@ message = 'Hello'
 names = ['World', 'Universe']
 items = [html('<li>{label}</li>') for label in names]
 result = render(
-    html('''
+    html(
+        '''
   <ul title="{message}">
     {items}
   </li>
-''')
+'''
+    )
 )
 expected = '<ul title="Hello"><li>World</li><li>Universe</li></ul>'